pyproject.toml
[tool.poetry]
name = "claude-chat-ui"
version = "0.1.0"
description = ""
authors = ["Jared Wuliger <hello@jaredwuliger.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.38.0"
anthropic = "^0.34.1"
pillow = "^10.4.0"
httpx = "^0.27.2"
google-auth = "^2.34.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-cloud-aiplatform = "^1.64.0"
black = "^24.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
